Add files via upload
My first implementation of the Game of Life. It works, but it's not optimised and it's pretty clunky. Run the file by typing the following into the terminal:

python gameoflife.py [width] [height] [square_size]

Where width, height and square_size are pixel measurements. A example might be:

python gameoflife.py 1000 800 20

The following controls are available:

- left click : create a live cell
- right click : create a dead cell
- 'f' : fills the entire board with live cells
- 'c' : clears entire board
- 'r' : reset the board to a previous state
- 'b' : print current board to console (more of a debugging thing)
- 'space' : switches  between editing and running mode
- 'q' : quit game